k/client: rename retry_with_mitigation to utils
(cherry picked from commit 217d25fc4ac5c72482444201154f62d1839ca464)
