+++ src/v/kafka/client/client.cc
@@ -15,8 +15,8 @@
 #include "kafka/client/exceptions.h"
 #include "kafka/client/logger.h"
 #include "kafka/client/partitioners.h"
-#include "kafka/client/retry_with_mitigation.h"
 #include "kafka/client/sasl_client.h"
+#include "kafka/client/utils.h"
 #include "kafka/protocol/errors.h"
 #include "kafka/protocol/fetch.h"
 #include "kafka/protocol/find_coordinator.h"
