+++ src/v/kafka/client/client.h
@@ -18,10 +18,10 @@
 #include "kafka/client/consumer.h"
 #include "kafka/client/fetcher.h"
 #include "kafka/client/producer.h"
-#include "kafka/client/retry_with_mitigation.h"
 #include "kafka/client/topic_cache.h"
 #include "kafka/client/transport.h"
 #include "kafka/client/types.h"
+#include "kafka/client/utils.h"
 #include "kafka/protocol/create_topics.h"
 #include "kafka/protocol/fetch.h"
 #include "kafka/protocol/list_offsets.h"
