+++ src/v/kafka/client/test/retry_with_mitigation.cc
@@ -7,7 +7,7 @@
 // the Business Source License, use of this software will be governed
 // by the Apache License, Version 2.0
 
-#include "kafka/client/retry_with_mitigation.h"
+#include "kafka/client/utils.h"
 
 #include <seastar/core/future.hh>
 #include <seastar/testing/thread_test_case.hh>
