+++ src/v/kafka/client/utils.h
